Create utils under tests and move the util modules into it
