+++ src/test/qa/dataset.js
@@ -6,7 +6,7 @@ const P = require('../../util/promise');
 const fs = require('fs');
 const dbg = require('../../util/debug_module')(__filename);
 const util = require('util');
-const s3ops = require('../qa/s3ops');
+const s3ops = require('../utils/s3ops');
 const readline = require('readline');
 const promise_utils = require('../../util/promise_utils');
 const argv = require('minimist')(process.argv);
