+++ src/test/qa/system_config.js
@@ -7,7 +7,7 @@ const net = require('net');
 const dgram = require('dgram');
 const http = require('http');
 const promise_utils = require('../../util/promise_utils');
-const server_ops = require('../qa/functions/server_functions');
+const server_ops = require('../utils/server_functions');
 const argv = require('minimist')(process.argv);
 const serverName = argv.server_ip;
 const api = require('../../api');
