+++ src/test/qa/upgrade_test.js
@@ -4,8 +4,8 @@
 var api = require('../../api');
 const P = require('../../util/promise');
 const promise_utils = require('../../util/promise_utils');
-const s3ops = require('../qa/s3ops');
-const ops = require('../system_tests/basic_server_ops');
+const s3ops = require('../utils/s3ops');
+const ops = require('../utils//basic_server_ops');
 var AzureFunctions = require('../../deploy/azureFunctions');
 var cloudCD = require('cloud-cd');
 var path = require('path');
