+++ src/test/system_tests/test_cloud_pools.js
@@ -8,7 +8,7 @@ var _ = require('lodash');
 var AWS = require('aws-sdk');
 var argv = require('minimist')(process.argv);
 var P = require('../../util/promise');
-var basic_server_ops = require('./basic_server_ops');
+var basic_server_ops = require('../utils/basic_server_ops');
 var dotenv = require('../../util/dotenv');
 dotenv.load();
 var promise_utils = require('../../util/promise_utils');
