+++ src/test/utils/vm-helper.js
